Do a full reboot after bootloader unlock
This works around the fact that the Nexus 5 does
a factory reset after unlocking the bootloader when
you reboot from recovery (animation of android with
ears twitching).

If Maru is flashed within recovery mode directly after
the unlock, PerspectiveService does not start properly
and mcprepare errors out saying it can't find the
jessie-rootfs tar on /userdata. I suspect that /userdata
is wiped and re-formatted.

The ideal solution is to have Maru stuff on a separate
partition but that will only work on custom hardware.

The workaround is to reboot without flashing Maru and let the
reset happen with the old ROM. Then re-run the script, this
time with the bootloader unlocked, and only then flash Maru.

Also some wording changes.