Introduce new cross-platform installer in Go
I decided to re-write the installers from scratch in Go to make them a lot
better and simplify the overall Maru release process.

Source code is the same across platforms so no more need to maintain
platform-specific scripts (yay!). Go makes cross-compilation a joy.

We have the ability to download files from the network in a
cross-platform way. That means we can decouple the installers from the
images; that makes it a lot easier to maintain since we don't need to
re-package new installers for every new release, only if there is an
update to the installer itself. This means that the installer version
numbers will no longer correspond to Maru release version numbers.

The door is now open to doing automated backups since we use TWRP to
flash the update zip. It would even be possible to offer gapps
restoration as part of the installer...

The end result is that the release process is vastly simplified with
these new installers. Each Maru release only needs to ship a single
update zip per device. The installer zips (three total, one per OS) only
need to be updated when the installer itself is updated--most of the
time the same installer can be used even for new Maru releases.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>