Fix release zip false positive for similar device names
It is possible that two devices have similar names; specifically, there
is the case where there is a supported device "hammerhead" and some new
device called "hammer" that is a subset of a supported device. If we are
not careful, we may try to use the hammerhead update zip for the hammer
device.

We avoid any chance of this happening by setting the match string to be
delimited by "-", so the device name must be an exact match.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>