Use an isolated working dir
On Mac, the working dir used when double-clicking the binary is the
user's home dir. That means the downloaded images get stored in the
user's home dir too, which would be a surprising side effect of running
the installer. Let's try to make sure we use the installer dir itself as
our working dir to avoid this.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>