Avoid delay before showing welcome message
The adb/fastboot sanity checks were taking a while to run (adb server
needs to be started) so there were 2-3 secs of a blank terminal when the
user first runs the installer.

Let's move the sanity checks to AFTER the welcome message is shown so
the user isn't confused (I was!).

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>