linux: Refresh linux installer scripts state
1. Upgrade to 28.0.1 to support A/B updates and other
bug fixing. Why not 28.0.2 ? Because it needs Windows updating,
if we update windows version to be consistency with Linux, what
is difficult for some user. So I just upgrade it to 28.0.1
to match our needs, and doesn't need Windows updating.

2. Normally, we use grep -i "unlocked" get correct unlocked
state of device, for example bullhead, because its firmware
only prints "Device unlocked: true/false". But sailfish also
prints "Device critical unlcoked: true/false", so origin shell
script will get "true\r\nunlocked:" when sailfish is unlocked.
In current checking statement, it will redirect to LOCKED always.
So I change grep command to grep -i "device unlocked", and test
it with sailfish and bullhead, and it works fine.

3. 51-android.rules can't be reached now, and in source.android.com,
google says user can get 51-android.rules by installing adb package,
which is matained by community. But I don't find the real path for it.
So I change it to a URL mantained by a github user.

Signed-off-by: utzcoz <utzcoz@outlook.com>